rename mathScope reducer to mathSymbols
This is more inline with the scope/symbols distinction in utilities/mathParser
