+++ src/containers/MathObjects/MathSymbols/Variable/actions.js
