+++ src/containers/MathObjects/MathSymbols/Variable/components/Variable.js
