+++ src/containers/MathObjects/MathSymbols/reducer.js
