+++ src/containers/MathObjects/index.js
@@ -1,8 +1,8 @@
 export { default as Point } from './MathGraphics/Point'
 export { createPoint } from './MathGraphics/Point/actions'
 
-export { default as Variable } from './MathScope/Variable'
-export { createVariable } from './MathScope/Variable/actions'
+export { default as Variable } from './MathSymbols/Variable'
+export { createVariable } from './MathSymbols/Variable/actions'
 
 export { default as Folder } from './Folder'
 export { createFolder } from './Folder/actions'
