+++ src/containers/MathObjects/mathObjectTypes.js
@@ -5,7 +5,7 @@ export const VARIABLE = 'VARIABLE'
 export const mapTypeToState = {
   [FOLDER]: 'folders',
   [POINT]: 'mathGraphics',
-  [VARIABLE]: 'mathScope'
+  [VARIABLE]: 'mathSymbols'
 }
 
 export const mathObjectTypes = Object.keys(mapTypeToState)
