+++ src/reducer.js
@@ -3,7 +3,7 @@ import drawers from './containers/Drawer/reducer'
 import sortableTree from './containers/SortableTree/reducer'
 import folders from './containers/MathObjects/Folder/reducer'
 import mathGraphics from './containers/MathObjects/MathGraphics/reducer'
-import mathScope from './containers/MathObjects/MathScope/reducer'
+import mathSymbols from './containers/MathObjects/MathSymbols/reducer'
 import activeObject from './containers/MathObjects/services/activeObject/reducer'
 
 export default combineReducers( {
@@ -11,6 +11,6 @@ export default combineReducers( {
   sortableTree,
   folders,
   mathGraphics,
-  mathScope,
+  mathSymbols,
   activeObject
 } )
